+++ plugin/src/main/resources/command.rdcml
