--- we-1_13/build.gradle
@@ -1,37 +1,0 @@
-plugins {
-    id 'java'
-}
-
-group 'me.untouchedodin0.privatemines'
-version '3.0.5-SNAPSHOT'
-sourceCompatibility = targetCompatibility = JavaVersion.VERSION_1_8
-
-repositories {
-    mavenCentral()
-    maven { url 'https://jitpack.io' }
-    maven { url = 'http://maven.sk89q.com/repo/' }
-    maven { url = "https://repo.codemc.org/repository/maven-public/" }
-    maven {
-        url = 'https://hub.spigotmc.org/nexus/content/repositories/snapshots/'
-
-        // As of Gradle 5.1, you can limit this to only those
-        // dependencies you expect from it
-        content {
-            includeGroup 'org.bukkit'
-            includeGroup 'org.spigotmc'
-        }
-    }
-}
-
-dependencies {
-    testImplementation 'org.junit.jupiter:junit-jupiter-api:5.6.0'
-    testRuntimeOnly 'org.junit.jupiter:junit-jupiter-engine'
-
-    compileOnly 'com.sk89q.worldedit:worldedit-bukkit:7.2.0-SNAPSHOT'
-    implementation 'org.codemc.worldguardwrapper:worldguardwrapper:1.2.0-SNAPSHOT'
-    implementation 'com.github.Redempt:RedLib:6.0.6'
-}
-
-test {
-    useJUnitPlatform()
-}
