--- we-1_8/build.gradle
@@ -1,39 +1,0 @@
-plugins {
-    id 'java'
-}
-
-group 'me.untouchedodin0.privatemines'
-version '3.0.5-SNAPSHOT'
-sourceCompatibility = targetCompatibility = JavaVersion.VERSION_1_8
-
-repositories {
-    mavenCentral()
-    maven { url 'https://jitpack.io' }
-    maven { url = 'https://maven.sk89q.com/repo/' }
-    maven {
-        url = 'https://hub.spigotmc.org/nexus/content/repositories/snapshots/'
-
-        // As of Gradle 5.1, you can limit this to only those
-        // dependencies you expect from it
-        content {
-            includeGroup 'org.bukkit'
-            includeGroup 'org.spigotmc'
-        }
-    }
-}
-
-dependencies {
-    testImplementation 'org.junit.jupiter:junit-jupiter-api:5.6.0'
-    testRuntimeOnly 'org.junit.jupiter:junit-jupiter-engine'
-
-    compileOnly 'org.spigotmc:spigot-api:1.17-R0.1-SNAPSHOT' // The Spigot API with no shadowing. Requires the OSS repo.
-    compileOnly("com.sk89q.worldedit:worldedit-bukkit:6.1.4-SNAPSHOT") {
-        exclude group: 'org.bukkit'
-    }
-    compileOnly "com.sk89q.worldguard:worldguard-legacy:6.1.3-SNAPSHOT"
-    implementation 'com.github.Redempt:RedLib:6.0.6'
-}
-
-test {
-    useJUnitPlatform()
-}
