+++ we_6/build.gradle
@@ -1,0 +1,4 @@
+dependencies {
+    compileOnly project(':compat')
+    compileOnly 'com.sk89q.worldguard:worldguard-legacy:6.1.3-SNAPSHOT'
+}
